poc/pyproject.toml
[project]
name = "dot-poc"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3",
    "mcp",
    "pyyaml",
]
